Moved getting sample analyses into a new service
